Moved classes to files.
